✨ feat(page): redesign home page to landing page with feature cards
- Replace resume builder form with clean landing page featuring AI analysis, resume builder, and recommendations
- Update app metadata to reflect Filecoin and 0g focus
- Remove deployment announcement step from CI workflow

BREAKING CHANGE: Home page UI completely redesigned, removing previous resume form functionality
