+++ .github/workflows/deploy-vercel.yml
@@ -403,24 +403,6 @@ jobs:
               // Don't fail the workflow if homepage update fails
             }
 
-      - name: "📢 Announce deployment and favorite frame"
-        if: github.ref == 'refs/heads/main' || github.ref == 'refs/heads/master'
-        env:
-          FARCASTER_BEARER_TOKEN: ${{ secrets.FARCASTER_BEARER_TOKEN }}
-        run: |
-              PRODUCTION_URL="${{ steps.deploy-production.outputs.production_url }}"
-              # Extract domain from full URL for scripts that expect domain only
-              DOMAIN=$(echo "$PRODUCTION_URL" | sed 's|https://||' | sed 's|http://||')
-              
-              echo "❤️ Debugging domain manifest for $DOMAIN"
-              node scripts/debug-domain-manifest.js "$DOMAIN"
-    
-              echo "❤️ Favoriting frame for $DOMAIN" 
-              node scripts/favorite-frame.js "$DOMAIN"
-              
-              echo "🚀 Posting to Farcaster about the new deployment"
-              node scripts/post-to-farcaster.js "New version deployed at $DOMAIN" "$PRODUCTION_URL"
-    
       - name: "🔍 Deploy to Vercel (Preview)"
         if: github.event_name == 'pull_request'
         id: deploy-preview
